selinux: Add support for changing enforcing state in kernel
SELinux is always a barrier for doing hacky tricks in kernel. I know
it's always not suggested to permissive selinux, but it's just far too
expensive for me to edit the policy.

Add an api to allow permissive selinux for a short amount of time. Let's
pray for our kernel to not get exploited in this grace period.

Adapted to Linux v5.10 kernel.

Signed-off-by: LibXZR <i@xzr.moe>
Signed-off-by: Levi Marvin <levimarvin@icloud.com>